[RenderDll][XRenderMetal] Enhance shader management and rendering capabilities
- Expanded shader management with new classes for uniform and texture bindings, improving flexibility and maintainability.
- Introduced additional uniform retrieval methods in `CMetalBaseRenderer` for better shader interaction.
- Enhanced blending state handling in the Metal pipeline, allowing for more complex rendering scenarios.
- Updated shader parsing logic to support new features and improved error handling.
- Added tests to validate the new shader management functionalities and ensure robustness in rendering processes.

+++ RenderDll/XRenderMetal/MetalStateCache.m
@@ -28,7 +28,7 @@ struct MetalPipelineStateKey
     uint64_t vertexFunctionHash;
     uint64_t fragmentFunctionHash;
     uint64_t vertexFormatHash;
-    uint32_t renderStateHash;
+    uint64_t renderStateHash;
     MTLPixelFormat colorPixelFormat;
     MTLPixelFormat depthPixelFormat;
     
@@ -50,7 +50,7 @@ struct MetalPipelineStateKeyHash
         size_t h1 = std::hash<uint64_t>{}(key.vertexFunctionHash);
         size_t h2 = std::hash<uint64_t>{}(key.fragmentFunctionHash);
         size_t h3 = std::hash<uint64_t>{}(key.vertexFormatHash);
-        size_t h4 = std::hash<uint32_t>{}(key.renderStateHash);
+        size_t h4 = std::hash<uint64_t>{}(key.renderStateHash);
         return h1 ^ (h2 << 1) ^ (h3 << 2) ^ (h4 << 3);
     }
 };
